backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo>=4.5,<4.7
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
cryptography==41.0.7
openai==1.6.1
anthropic==0.8.1
itsdangerous==2.1.2
